tut: Use different prebreak and postbreak symbols.
For prebreak, use a backslash. Conveniently, the backslash is also used
for continuations on the command line, so copying it should not be an
issue.

For postbreak, use a different arrow. Also, colour it red to stand out
better. This symbol also seems to be non-copyable, maybe because of the
rotatebox, but being an extraneous symbol, we probably want it that
way anyway.